+++ NEWS.md
@@ -1,4 +1,5 @@
-+ added a error message to remind that "sequences must have unique names".
++ added option to show the fill legend.(2021-03-23, Tue)
++ added a error message to remind that "sequences must have unique names".(2021-03-18, Thu)
 + added `ggSeqBundle()` to plot Sequence Bundles for MSAs based `ggolot2` (2021-03-18, Thu)
 
 # ggmsa 0.0.6